EMAILPLT-95, EMAILPLT-13
Partial fix.  The Encryption Service now adds an exception that suggests checking to see if the encryption password is set or had changed in the ApplicationContext.xml file.  Also added warning to not change encryption password to applicationContext.xml.
Ideally still need to 1) indicate to admin password is not set when EmailPortlet is configured, and 2) give user some ability to recover from password decryption failure